more styling for timer.js
