+++ client/src/App.css
@@ -6,11 +6,12 @@ html {
     margin: 0;
     padding: 0;
     box-sizing: inherit;
+    /* background-color: rgb(53, 53, 53); */
     background-color: black;
     /* background-color: white; */
+    width: 100%;
 }
 html, body {
-    width: 100%;
     height: 100%;
     /* color: rgb(189, 189, 189); */
     color: lime;
@@ -21,40 +22,38 @@ html, body {
     display: flex;
     justify-content: center;
 }
-
-/* ===== =====  HEADINGS  ===== ===== */
-h1 {
-    font-size: 170%;
-    margin-top: 20px;
-    margin-bottom: 30px;
-    /* color: lime; */
-    letter-spacing: 1px;
-}
-h2 {
-    justify-content: center;
-    color: lime;
-    font-size: 120%;
-
-}
 .main {
     display: flex; 
     flex-direction: column;
     justify-content: center;
     align-items: center;
+    width: 100%;
+}
+
+/* ===== =====  HEADINGS  ===== ===== */
+h1 {
+    text-align: center;
+    font-size: 170%;
+    letter-spacing: 1px;
+    margin: 20px 0 10px;
+}
+h2 {
+    justify-content: center;
+    color: lime;
+    font-size: 120%;
 }
 
 /* ===== =====  REUSABLE  ===== ===== */
 .btn:link,
 .btn:visited,
 input[type=submit] {
+    width: auto;
     padding: 3px 8px;
     font-size: 70%;
     font-weight: 300;
     border-radius: 200px;
-    /* display: flex; */
-    justify-content: center;
-    align-content: center;
     color: white;
+    margin-bottom: 5px;
     background-color: rgb(39, 39, 39);
     transition: background-color 0.2s, border 0.2s, color 0.2s;
 }
@@ -79,14 +78,17 @@ input[type=submit]:active {
 .links:link,
 .links:visited {
     text-decoration: none;
-    margin-right: 5px;
+    width: auto;
     font-size: 80%;
     color: lime;
     word-spacing: 2px;
     letter-spacing: 0.5px;
-    padding-bottom: 1px;
+    padding-bottom: 5px;
+    margin-left: 10px;
     border-bottom: 2px solid transparent;
     transition: border-bottom 0.3s, color 0.3s;
+    display: flex;
+    justify-content: center;
 }
 .links:active,
 .links:hover {
@@ -95,8 +97,8 @@ input[type=submit]:active {
 
 }
 /* ===== =====  FORMS  ===== ===== */
-.title {
-    display: flex;
+.title h2{
+    display: flex; 
     justify-content: center;
     margin-bottom: 15px; 
 }
@@ -145,7 +147,7 @@ label {
     display: flex;
     flex-direction: column;
     justify-content: center;
-    /* align-content: center; */
+    align-items: center;
 }
 /* ===== =====  REGISTER  ===== ===== */
 .section-register {
@@ -161,16 +163,36 @@ label {
 .navbar {
     display: flex; 
     justify-content: flex-end;
-    /* justify-content: space-around; */
-    padding: 5px;
-    margin-bottom: 5px;
-    /* border: 1px solid white; */ 
+    align-content: center;
+    padding: 10px;
+    margin: 10px 0;
+    /* border: 1px solid white;  */
 }
 
 /*  ===== =====  TIMER  ===== ===== */
 .timer {
     border: 2px solid white;
+    width: auto;
+    display: flex;
+    /* justify-content: center; */
+    flex-direction: column;
+    /* align-content: center; */
 }
+p {
+    font-size: 70%;
+    text-align: center;
+}
+
+.App-clock {
+    transform: translateX(-50%) translateY(-50%);
+    color: lime;
+    font-size: 170%;
+    font-family: Orbitron;
+    letter-spacing: 7px;
+    text-align: center;
+    width: 50%;
+}
+
 /* ===== =====  AVATAR  ===== ===== */
 .avatarImg {
     width: 30px;
@@ -186,14 +208,3 @@ td {
 
 
 
-/* 
-  .App-clock {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translateX(-50%) translateY(-50%);
-  color: lime;
-  font-size: 70px;
-  font-family: Orbitron;
-  letter-spacing: 7px;
-} */
