+++ client/src/MyAccount.js
@@ -41,7 +41,11 @@ class MyAccount extends Component {
         return (
             <div>
                 <h2>MyAccount</h2>
-                <Link to = {{pathname: '/settings', state: {thisUser: theUser}}} className='links'>
+                <Link to = {{
+                    pathname: '/settings', 
+                    state: {
+                        thisUser: theUser
+                        }}} className='links'>
                    Settings
                 </Link>
